Rewritten using FFI instead of being a compiled extension.
I've tested this works with Ruby 1.8, Ruby 1.9 and JRuby 1.2.

I've taken the opportunity to fix some problems with the existing
extension: now namespace the Password class with Cracklib; can now pass
a dictpath to the #strong? and #weak? methods, or omit them to use the
default that rubylibcrack tries to find.